Reorganize workspace routes tests
